fix: user profile picture upload

+++ src/main/java/ru/mirea/kotiki/repositories/UserRepository.java
@@ -10,7 +10,7 @@ public interface UserRepository extends ReactiveCrudRepository<User, Long> {
 
     Mono<User> findByEmail(String email);
 
-    Mono<User> findById(Integer id);
+    Mono<User> findById(Long id);
 
     Mono<Boolean> existsByEmail(String email);
